feat:role wise permission

+++ app/Http/Controllers/RoleController.php
@@ -3,7 +3,9 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Spatie\Permission\Models\Permission;
 use Spatie\Permission\Models\Role;
+
 
 class RoleController extends Controller
 {
@@ -23,7 +25,8 @@ class RoleController extends Controller
     public function create()
     {
 
-        return view('role-permission.role.create');
+        $permissions = Permission::all();
+        return view('role-permission.role.create', ['permissions' => $permissions]);
     }
 
     /**
@@ -32,17 +35,37 @@ class RoleController extends Controller
     public function store(Request $request, Role $role)
     {
         $request->validate([
-            'name' => [
-                'required',
-                'string',
-                'unique:permissions,name',
-            ],
-
+            'name' => 'required|string|unique:roles,name'
         ]);
 
-        $role->name = $request->input('name');
-        $role->save();
 
+        $role->name = $request->input('name');
+        $role->givePermissionTo($request->input('permission'));
+
+        # check if the role already exists
+        // if ($role->exists) {
+        //     return redirect()->back()->with('error', 'Role already exists');
+        //     }
+        //     # save the role
+        //     // $role->save();
+
+        //     # assign the permissions to the role
+        //     $role->givePermissionTo($request->input('permission'));
+        //     return redirect()->route('role.index')->with('success', 'Role created successfully');
+
+
+        # assign the permissions to the role
+        // if(!empty($request->permission)){
+        //     // $role->syncPermissions($request->permission);
+        //     foreach ($request->permission as $name) {
+        //         $role->givePermissionTo($name);
+        //     }
+        // }
+        // else{
+        //     return redirect()->back()->with('error', 'Please select at least one permission');
+        // }
+
+        $role->save();
         return redirect('role')->with('success', 'Role Created Successfully');
     }
 
@@ -60,7 +83,14 @@ class RoleController extends Controller
     public function edit(Role $role)
     {
 
-        return view('role-permission.role.edit', ['role' => $role]);
+        $permissions = Permission::all();
+        # get permission on role
+        $hasPermissions = $role->permissions->pluck('name');
+        return view('role-permission.role.edit', [
+            'role' => $role,
+            'hasPermissions' => $hasPermissions,
+            'permissions' => $permissions,
+        ]);
     }
 
     /**
@@ -81,6 +111,18 @@ class RoleController extends Controller
             'name' => $request->name
         ]);
 
+        if (!empty($request->permission)) {
+            // $role->givePermissionTo($request->permission);
+
+            $role->syncPermissions($request->permission);
+        } else {
+            // $role->permissions()->detach();
+            $role->syncPermissions([]);
+        }
+
+        // $role->syncPermissions($request->permission);
+        // $role->givePermissionTo($request->input('permission'));
+
         return redirect('role')->with('success', 'Role Updated Successfully');
     }
 
@@ -90,7 +132,57 @@ class RoleController extends Controller
     public function destroy(Role $role)
     {
 
+        if ($role == null) {
+            session()->flash('error', 'Role not Found');
+            return response()->json(['status' => false]);
+        }
+
+
+        // $role->permissions()->detach($role);
+
+
+        // $permission->removeRole($role);
+        $permission = Permission::get();
+        // $permission->detach($role);
+        $role->revokePermissionTo($permission);
         $role->delete();
+
+        session()->flash('success', 'Role Deleted successfully');
+
+        // return response()->json(['status' => true]);
+
         return redirect('role')->with('success', 'Role Deleted Successfully');
+    }
+
+    public function addPermissionRole(Role $role)
+    {
+
+        $permissions = Permission::get();
+        return view(
+            'role-permission.role.add-permission',
+            [
+                'role' => $role,
+                'permissions' => $permissions
+            ]
+        );
+    }
+    public function givePermissionRole(Request $request, $role)
+    {
+
+        // validation
+        $request->validate([
+            'permission' => 'required'
+        ]);
+
+
+        $roles = Role::findOrFail($role);
+        // return $roles;
+        // $roles->permissions()->sync($request->permission);
+        $roles->syncPermissions($request->permission);
+
+
+        return redirect('role')->with('success', 'Permission added to Role');
+        // return redirect()->back()->with('status', 'Permission added to Role');
+
     }
 }
